[UPG][3756241] school_lunch: improve models grouping (#3)
- The Python models of this custom module are regrouped in separate files that better highlight the group of models they belong to
