+++ school_lunch/models/__init__.py
@@ -1,5 +1,8 @@
-from . import models
-from . import loyalty_card
-from . import sale_order_line
-from . import res_config_settings
 from . import ir_http
+from . import loyalty_card
+from . import product_product
+from . import res_config_settings
+from . import res_company
+from . import res_partner
+from . import sale_order
+from . import school_lunch
