+++ school_lunch/models/sale_order.py
@@ -47,14 +47,3 @@ class SaleOrder(models.Model):
                 if set_qty or (add_qty is not None):
                     return {"line_id": line.id, "quantity": line.product_uom_qty, "option_ids": []}
         return super()._cart_update(product_id, line_id, add_qty, set_qty, *args, **kwargs)
-
-
-class ProductProduct(models.Model):
-    _inherit = "product.product"
-
-    def _is_add_to_cart_allowed(self):
-        self.ensure_one()
-        for mt in ("soup", "meal"):
-            if self.id == self.env.ref("school_lunch.product_" + mt).id:
-                return True
-        return super()._is_add_to_cart_allowed()
